diff/push: normalize utf8mb3 across shards w/ different server versions
In a sharded environment, it is possible for some shards to have different
server versions. The legacy 3-byte utf8 character set is represented as
"utf8mb3" in recent versions (MySQL 8.0.29+, MariaDB 10.6+); ditto for its
collations as "utf8mb3_..." (MySQL 8.0.30+, MariaDB 10.6+).

This commit improves diff logic to be aware of this fact when running diff or
push against multiple servers of differing versions. This way, extraneous
ALTER statements/clauses are not generated when tables/columns only differ due
to their server version's representation of the legacy 3-byte utf8 character
set or its collations.

Users can override this behavior with the --exact-match option, as with other
similar purely-cosmetic table differences.