Golang: bump to 1.18 for CI, modules
Now that Go 1.19 is available, the oldest Go version supported by the Go
authors is 1.18. Skeema follows the same logic for minimum Go version,
so this commit bumps go.mod and CI to specify 1.18.

The Skeema codebase may start to use Go 1.18 features in the near future,
for example Generics, strings.Cut(), and debug.ReadBuildInfo().

Note that the release pipeline (powered by GoReleaser) always uses the latest
Go for building binary releases, so no change is necessary there.

This commit also updates the version of github.com/skeema/mybase to be on a
tagged release which also includes the corresponding Go version change.