workspace: fix nil pointer panic if unstable network connection to DB
If Skeema is able to connect to the workspace initially, but a subsequent
connection attempt fails, a panic would result due to a logic bug in a
deferred error handler. Realistically this would only arise in rare situations
where the network connection to the database is unstable/flaky or extremely
high-latency.

This commit fixes the logic bug, and also cleans up a few other related call
patterns in the same package. The cause was a pattern of having a deferred
anonymous function manipulate named return values during error conditions.
Error conditions typically return nil *something, non-nil error; but the
deferred error handler needs *something to be non-nil as well in order for its
logic to work properly. This is too brittle, so an alternative approach is now
used.