password prompt: prefer STDERR for the interactive prompt text
When using the -p (--password) option with no value supplied, Skeema prompts
for the password interactively, just like the mysql client. Previously the
"Enter password:" prompt was printed out STDOUT, since this is where the mysql
client prints it.

However, it is not unusual to redirect Skeema's STDOUT to a file or pipe to
another program, since all other Skeema STDOUT output is designed to be valid
SQL. When combining STDOUT redirection with -p, this presents two problems:

1. Skeema may confusingly appear to be hung, since the "Enter password:"
   prompt was redirected too; see #188

2. The first line of redirected output includes invalid SQL, since "Enter
   password:" isn't valid SQL.

This commit modifies the password prompt text to prefer STDERR. Only one case
remains where STDOUT is used instead: if STDERR is *not* a terminal, but
STDOUT *is* a terminal, then STDOUT is used. This is designed to help ensure
the prompt text is visible on the terminal.

Improves #188.