internal refactoring / minor codebase alignment (2022 Q3)
This commit brings in some minor internal changes to reduce the number of
unnecessary divergences between the open source Community edition and closed
source Premium edition codebases.

This is being done ahead of plans to move a few of Premium's smaller features
into Community for Skeema v1.9 -- more on that to come in future commits. No
real functional impact yet in this commit.