password option: support separate STDIN prompting per dir/host
Background:
To match the behavior of the standard MySQL client, Skeema's password option
supports the ability to prompt for the password on STDIN if the option is
supplied in a way that explicitly lacks a value on the command line or option
file. Previously, this prompt-on-STDIN behavior required either supplying
--password on the CLI without a value, or having a bare "password" line (with
no equals sign or value) in a *global* option file.

Behavior change in this commit:
This commit permits handling of a bare "password" line to be processed
properly in *any* option file, not just a global one. Skeema commands,
which generally execute recursively from the working directory, will now
prompt for the password whenever operating on a directory whose .skeema file
contains a bare "password" line. The prompt text will indicate which DB host
the password is for (or which directory, if the host is not specified yet).

Once the password is entered, this value will then apply to all subdirs of
that directory as well, unless those directories have a .skeema file which
overrides the password option again in turn.

Practical usage:
When operating on multiple database instances at once (e.g. `skeema diff` from
the top of a schema repo containing distinct schemas for different hosts), it
is now possible to put bare "password" lines in each of the .skeema files at
the host level, and Skeema will prompt for separate passwords for each host.
Previously, this was not possible without placing the plaintext password
values into the .skeema files, which is extremely inadvisable for security
reasons.

Related future directions:
Subsequent commits will add the ability to use arbitrary ENV variables for
options, including the password option. This will provide a non-interactive
alternative for using different passwords for each host, again without
requiring raw passwords in .skeema files.