option files: support environment variables for some option values
This commit adds the ability to use environment variables in option files,
meaning .skeema files as well as "global" option files. For example, the line
"password=$MY_SKEEMA_PASSWORD" will dynamically use the value of the
environment variable MY_SKEEMA_PASSWORD for the password option.

So far, the following options support using environment variables for values:

* user
* password
* host
* port
* socket
* schema
* temp-schema

In the case of host and schema, if the environment variable value is a comma-
separated list of multiple values, this will properly expand to handle sharded
situations exactly as if the comma-separated list had appeared directly in the
.skeema file.

Environment variable substitution only works for full values. In other words,
usage such as "schema=mydb_$FOO" or "host=${DBSUBDOMAIN}.myhost.com" is not
supported at this time.

If the value is wrapped in single-quotes, it will not be processed as an
environment variable. This provides a mechanism to express values that just
happen to begin with a $ symbol but are not environment variables.