password option: avoid redundant prompting for same user@host
Recent commit ab4a84c added support for per-subdir interactive password
prompting, by way of having a "password" line with no =value in a subdir-level
.skeema file. The original logic memoized passwords per directory hierarchy
(avoiding redundant prompting for subdirs, if the parent dir has the bare
"password" configuration), but it contained no memoization for sibling subdirs
that have a bare "password" configuration and emit the same user and host.

This commit adjusts the logic to account for this, and also refactors the call
pattern to be more natural: individual Skeema subcommands no longer need to
check for directories which require password prompting. This allows many of
the changes from ab4a84c to be reverted while offering improved functionality.