password option: fix multi-host prompt text
Recent commit ab4a84c added support for per-subdir interactive password
prompting, by way of having a "password" line with no =value in a subdir-level
.skeema file. When a directory maps to more than 2 hosts, the prompt message
lists the first host and omits the others in favor of a host count. However,
this count had an off-by-one error. This commit fixes the prompt message to
accurately reflect the number of hosts that the dir maps to.

This is only a textual change with no impact on behavior.

This commit also updates module github.com/skeema/mybase to its latest commit,
although the new logic there is not used yet.