MariaDB: CREATE OR REPLACE syntax for procs/funcs in Community Edition
When modifying an existing routine in MariaDB 10.1+, `skeema diff` and
`skeema push` will now automatically use CREATE OR REPLACE syntax (instead
of emitting a DROP followed by a re-CREATE). Since it is atomic, this
operation does not require --allow-unsafe in MariaDB.

This functionality was already present in Skeema Premium CLI (v1.6.0+) and
Skeema Cloud Linter since Nov 2021. This commit brings it into Skeema
Community CLI.

Note that the OR REPLACE syntax for routines is not available in MySQL or
Percona Server, so for those flavors, modifications to existing routines
still use the separate non-atomic DROP / re-CREATE flow, requiring use of
the --allow-unsafe option.