internal: align old SQL processor with new one
This commit changes some minor behaviors of the old SQL file processor, to
more closely align with the new (not-yet-merged) parser, in terms of how some
edge cases of Statement fields are handled. This will make it easier to re-
use the testing logic while temporarily supporting both parsers at once in
the future.

* Statement.nameClause now omits trailing whitespace/comments

* fs tests now compare Statement values more directly, since everything except
  the Error field is a simple scalar

* Even if a USE command is terminated with just a newline, Statement.Delimiter
  is still populated with the current client-side delimiter string

* fs.ParseStatementsInFile no longer takes a second arg with the initial
  delimiter (it's always ";" in all callsites anyway)

* Unsupported statements no longer populate Statement.Error with an error
  value; in this situation this value wasn't used/read anywhere