internal refactor: new SQL file processor
This commit completely replaces Skeema's previous regex-based SQL processing
logic with a new hand-written parser. In real-world testing, the new
implementation provides an 8x to 40x speedup on parsing time, while also using
substantially less memory.

Background:

Skeema largely does not rely on fully parsing SQL to understand table
structures, and this will remain the case. However, it needs to parse enough
of SQL to understand types of statements and names of objects, in order to
map the individual statements within *.sql files to database objects.

The previous code for this relied on outdated version of an external module,
github.com/alecthomas/participle, to implement a simple regex-based lexer and
parser. This generally worked fine, and a subdirectory containing several
hundred CREATE statements could be parsed in under 1 second on modern
hardware. However, if the directory also contained large INSERT statements
(for example from mysqldump), parsing time would be prohibitively slow, even
though Skeema does not interact with INSERT statements at this time.

Out-of-the-box SQL parser libraries aren't a great fit for Skeema's use-case
though, because Skeema intentionally needs to be permissive/fuzzy in its SQL
handling with respect to syntax errors. If you edit the definition for table
`foo` and accidentally introduce a syntax error typo in its CREATE, Skeema
understands that your intention was to alter `foo` but you just have a typo,
as opposed to wanting to drop `foo` since it no longer has a valid definition.

This commit:

The old lexer and parser are replaced with entirely new code, and the
participle dependency is removed from the module list.

Some related code in the fs and tengo modules are also affected, due to some
of the error semantics/conditions changing. This will likely be tweaked
further in upcoming commits.

Related issues:

See #22 and #190. Both remain open, but this commit gets closer to solutions
for both.