init: avoid parsing SQL files in working directory
Previously, `skeema init` would needlessly parse the *.sql contents of the
working directory, simply due to some underlying code being unnecessarily
coupled: the "create a subdirectory" logic was a method on the fs.Dir type,
so the working directory was parsed to get a fs.Dir before creating the host-
level subdirectory.

Ordinarily this was harmless, but if a very large *.sql file (such as a
logical dump from `mysqldump`) happened to be present in the working
directory, `skeema init` could appear hung and use lots of CPU while
processing the large *.sql file or files.

This situation was already recently improved by the new SQL processor in
commit e625c97, providing a 8x to 40x speedup in processing time. However,
a very large SQL file dump (hundreds of GB or more) could still cause a
noticeable stall in `skeema init`.

This commit solves the problem entirely by decoupling the subdir creation
logic from the other dir logic, so that the working directory is no longer
parsed. Fixes #190.