internal refactor to applier subpackage abstractions
This commit adjusts several of the abstractions in the applier subpackage, to
move away from being DDL-specific. Initial DML-related logic is being
introduced into Skeema Premium for the seed data feature. Some of that
functionality may eventually be added to the Community Edition in a limited
form in the future.