improve panic handler and related logging
This commit improves the handling of Golang panics in Skeema in several ways:

The primary error logging in a panic situation is now more descriptive, always
including the function/file/line, even if --debug is not enabled. The message
also describes more clearly that a panic has occurred, and how to report a bug
in the Skeema issue tracker.

Panics that occurred in the worker goroutines of diff/push previously were not
handled by the same panic handler as the main goroutine, potentially resulting
in truncated stack traces. This is now fixed, although goroutines from
subpackages still have the same deficiency for now.

Test coverage for panic and exit code related functions have been improved.