ci: bump Go version to 1.19 and actions to v3
Now that Go 1.20 is available, the oldest Go version supported by the Go
authors is 1.19. Skeema follows the same logic for minimum Go version,
so this commit bumps go.mod and CI to specify 1.19.

Note that the release pipeline (powered by GoReleaser) always uses the latest
Go for building binary releases, so no change is necessary there.

This commit also updates the GitHub Actions versions for actions/setup-go and
actions/checkout to v3, instead of v2. This prevents the deprecation warning
about Node.js 12 in the Actions workflow pages.