internal refactor: delimiter handling for output and dumping
This commit refactors how statement delimiters are handled in output logic
and SQL file logic. Detection of stored programs with compound statements
(requiring special DELIMITER handling) is now unified into a single code path
which does not rely on regular expressions.

Two minor functional changes are present:

When `skeema diff` or `skeema push` operates on multiple compound statements
in a row, the output no longer needlessly includes extraneous DELIMITER
commands in between each compound statement (restoring the standard ";"
delimiter and then immediately going back to "//" ahead of the next proc/
func).

When a hand-written .sql file contains a compound statement without being
preceded by a DELIMITER command, the parser now assumes the rest of the file
is part of the compound statement, as this tends to be true in most cases; and
when it is not true, an obvious error will result once the statement is used
in a workspace. Previously, the parser would emit StatementTypeUnknown
statements in this situation, which only display a warning.

Several structural code changes are also present:

tengo.Statement and tengo.RoutineDiff are both able to self-identify when
they represent a compound statement, via a common interface.

applier.Printer implementations are now responsible for adding proper
delimiters (and DELIMITER commands) to statement output as needed. The
applier.ClientState abstraction now includes information on whether a
statement needs a standard delimiter, special delimiter (stored program with
compound statement), or no delimiter (e.g. shell-out from alter-wrapper).

In the dumper subpackage, the method for adding a statement to a SQLFile has
been made more flexible, to not be CREATE specific. This change is necessary
ahead of the INSERT support for seed data functionality being added to Skeema
Premium.