diff/push: permit supported ALTERs on unsupported tables
Skeema does not support diffing tables that use some specific edge-case
features, as per https://www.skeema.io/docs/requirements/ (see "Unsupported
for ALTER TABLE" section). Previously, any use of those features completely
prevented Skeema from generating or running an ALTER TABLE for that table.

However, in some cases, the true diff may just represent a *supported*
alteration to an unsupported table: for example, adding a column or index to
a table that just happens to use sub-partitioning. This commit leverages
Skeema's existing diff verification guardrails to detect such situations, and
allow an ALTER on the table to proceed IF the generated SQL can be empirically
confirmed to be correct in a workspace.

This logic covers two categories of alter operations, which now can proceed as
of this commit:

* Making a supported change to a table which happens to use unsupported
  features, but those features are not impacted by the alteration

* Removing an unsupported feature from a table, such as dropping a spatial
  index

This logic is run even if the --verify option has been disabled. In that case,
regular verification of supported tables is skipped as requested, and only
unsupported table diffs are verified.

In some cases in MySQL 8, the alter may still fail verification and remain
marked as unsupported, due to some edge-cases in how SHOW CREATE TABLE handles
non-default collations.