internal: refactor applier.VerifyDiff to use options struct
This commit adjusts the arg signature for applier.VerifyDiff to use a new
VerifierOptions struct, instead of requiring an applier.Target be passed in
directly. This change makes the code more reusable in the future, if any
other subpackages ever need to use the "verification of unsupported diffs"
logic introduced in previous commit 7b5e86d.

This commit also adjusts an integration test, which was modified in the
previous commit in a way which inadvertently reduced code coverage slightly.
The affected lines are now covered again properly.