global options: minor tweaks to host, schema, port
Most commands do not allow host or schema to be provided on the command-line,
but the error message about this is rather succinct. This commit adds a link
to the online documentation explaining this in more depth.

This commit also improves --port so that its shortcut option -P may be used
for any command. Previously the shortcut -P only worked for `skeema init` and
`skeema add-environment`, but this was inconsistent with the --socket shortcut
of -S working for all commands.