tests: close connection pools in more integration tests
Some integration tests in the tengo subpackage instantiate Instance values
that are separate from the normal path used for every test. This commit
ensures that any cached connection pools for those Instances are closed
properly, since the normal before-test automatic handling for this won't
take care of it.

This is unlikely to affect test execution in any real way, since test
containers are shut down after each subpackage, but it's harmless to add.