internal: move util.ShellOut to its own package and refactor
This commit moves the util.ShellOut type (and associated functions) into its
own separate internal subpackage. The type and functions have been refactored
in the process, so this commit also adjusts all call sites accordingly.

This work is being done so that tengo can import shellout for later use in
Docker CLI calls. Previously this was not possible with shellout being in
util, as util depends on tengo.

This commit also enhances the internal shellout functionality in several ways:

* Shellouts may now have custom STDERR and/or STDIN
* A new method which captures STDOUT and STDERR separately has been added
* Variable interpolation now ignores shell/env variables of the form ${FOO},
  as well as Golang template expressions such as {{json .}}; previously these
  cases would throw an error