diff/push: add new option lax-column-order
This commit adds the ability to ignore differences in column ordering when
using `skeema diff` or `skeema push`. This behavior is controlled by new
option lax-column-order, which isn't enabled by default. When enabled, MODIFY
COLUMN clauses are suppressed whenever the only difference in the column is
the column's ordering within the table.

This option is intended for use in situations where some environments have
different column ordering, caused by historical reasons / technical debt.

This option does not affect MODIFY COLUMN clauses when the diff contains
additional changes to the column besides just the ordering. In this situation,
the entire MODIFY COLUMN clause (including the FIRST or AFTER part) is still
generated as-is. In other words, even with lax-column-order enabled, columns
will still be re-ordered if some other aspect of the column is also being
modified at the same time.

This option does not affect ADD COLUMN clauses. Any FIRST or AFTER portion of
an ADD COLUMN clause will still be present. In other words, ordering for new
columns is still respected.

Closes #178.