MySQL 8.1 support
This commit marks MySQL 8.1 as a supported DB flavor. MySQL 8.1 was released
on July 18 and is the first innovation (feature) release following MySQL's new
versioning scheme.

None of MySQL 8.1's changes directly impact DDL or schema management, however
some modifications were necessary to support Skeema's Docker-related
touch-points. MySQL 8.1 is not available on Dockerhub yet, which is unusual
since the release was two weeks ago -- especially as MySQL 8.0.34 *is*
available there, and that was released on the same day as 8.1.

To solve this problem, this commit adds logic to obtain MySQL 8.1+ community
server images from Oracle Container Registry instead. This affects Skeema's
workspace=docker feature, as well as the integration suite runners, whenever
a "mysql:8.1" image is specifically requested/configured in Skeema's flavor
option or in the SKEEMA_TEST_IMAGES env variable. In all other situations,
images are still obtained from Dockerhub as before.