MySQL 8.1: obtain Docker images from DockerHub
Images for MySQL 8.1.0 are finally available on DockerHub, 15 days after
8.1's release, and unfortunately just a few hours after the changes we made
in 23520ea.

So this commit adjusts Docker-related touch-points (workspace=docker, and
integration test suites) to download 8.1 from DockerHub like previous
releases, reverting the temporary logic for using Oracle Container Registry.

Some of the other image location flexibility code from 23520ea remains in
place and could be useful in the future. It may make sense to formalize this
logic into a hard-coded map in a future commit, so that the locations can be
adjusted more easily.