internal: file cleanups for introspection and diff logic
This commit moves the location of internal logic relating to introspection
and diffs. Previously, logic handling multiple object types (tables vs
routines) was often in the same file. This commit adjusts the locations as
follows:

* All routine (proc/func) introspection and diff logic is now in routine.go
* Table introspection logic is now in table_introspect.go
* Table diff logic is now in table_diff.go, except for alter clause logic
* Table diff alter clause logic is now in table_alter_clause.go
* The remaining code in diff.go is largely not object-type-specific.

No functionality changes are present in this commit.

This change is motivated by some upcoming minor improvements to diff handling
for routines. Previously, this was difficult to work with, since the routine
logic was spread among multiple files.