diff/push: support ALTER PROCEDURE and ALTER FUNCTION
For some limited types of adjustments to a routine, MySQL and MariaDB both
support ALTER statements to modify the routine's characteristics in-place.
Previously, Skeema would never emit ALTER PROCEDURE or ALTER FUNCTION
statements in these cases; all modifications to routines were handled with
a DROP and CREATE pair in MySQL, or a single CREATE OR REPLACE in MariaDB.

This commit adds the ability for `skeema diff` and `skeema push` to generate
an ALTER statement whenever the desired modification to a routine can be
handled solely with such a statement. In other words, an ALTER will now be
generated for changes that only affect some combination of a routine's SQL
characteristic, SECURITY characteristic, or comment clause.

Such changes are always considered safe (not requiring --allow-unsafe even in
MySQL) because the change is atomic, and does not cause the routine to briefly
not exist.

One notable special case is an attempt to *remove* a COMMENT clause from a
routine. A server bug in MySQL 8.0 and 8.1 prevents this from being possible
in an ALTER statement; such a statement (e.g. ALTER PROCEDURE foo COMMENT '')
simply does not remove any preexisting comment. To work-around this server
bug, Skeema will replace the routine instead of using an ALTER in this
specific situation.