diff/push: modifying proc/func params or return type is unsafe
This commit marks proc/func modifications as unsafe if the new version of the
routine changes the parameters or return type. This operation now requires
enabling --allow-unsafe to proceed, even when using MariaDB's atomic CREATE
OR REPLACE syntax.

This type of modification is risky because it may break call-sites in
application queries, and/or in other routines. There is no way to
simultaneously deploy application and routine changes in an atomic fashion,
and likewise for deploying modifications to multiple routines at once. This
means if the routine is already in-use anywhere, adjusting its parameters or
return type can cause query failures until all call sites have also been
adjusted.

Previously, modifying an existing routine was always considered safe in
MariaDB (Skeema automatically uses CREATE OR REPLACE), and always considered
unsafe in MySQL (which lacks CREATE OR REPLACE syntax). The latter remains the
case.