MariaDB 11.1 support
This commit marks MariaDB 11.1 as a supported DB flavor, now that its GA
release occurred on August 21. No functional changes were required, as none
of MariaDB 11.1's documented changes [1] directly affect Skeema's
functionality, and all integration test suites pass when run against a
MariaDB 11.1.2 Docker image.

At the moment, users of workspace=docker cannot use flavor=mariadb:11.1 yet
though, simply because the "11.1" tag unexpectedly does not exist yet on
DockerHub. I believe this is simply an oversight and have filed a GitHub issue
with MariaDB [2]. Since it is likely to be corrected in the near future /
prior to the next Skeema release, it isn't treated as a blocker for marking
11.1 as supported here.

[1] https://mariadb.com/kb/en/changes-improvements-in-mariadb-11-1/

[2] https://github.com/MariaDB/mariadb-docker/issues/529