ci, go.mod: bump Go version to 1.20
Now that Golang 1.21 is available, the oldest Go version supported by the Go
authors is 1.20. Skeema follows the same logic for minimum Go version, so
this commit bumps go.mod and CI to specify 1.20.

Note that the release pipeline (powered by GoReleaser) always uses the latest
Go for building binary releases, so no change is necessary there.

No Golang 1.20 features are used quite yet, but among them, it is possible
that multi-error wrapping (fmt.Errorf with multiple %w, and/or Unwrap
returning []error) may be used in the near future. And it is *very* likely
that quite a few 1.21 features will be used in 6 months, once Go 1.22 is
available.