workspace=docker: improve error if no docker CLI in PATH
Prior to Skeema v1.11.0, a built-in Docker client library was used for
workspace=docker, but this was removed in commit 2a2d8e3. This change can be
disruptive for users who enable workspace=docker while also running `skeema`
itself from a container environment. This is expected to affect only a small
minority of Skeema users, but nonetheless we should minimize the amount of
disruption by pointing such users to relevant documentation.

This commit improves the error message for `skeema diff`, `skeema push`, and
`skeema lint` when workspace=docker is enabled but `docker` is not found on
the PATH.

This commit will likely be reverted in 1 to 2 years, as it will become less
relevant as more users upgrade to newer Skeema versions.

Also note that although `skeema format` and `skeema pull --skip-format` also
use workspaces, their error messages are not adjusted here, since these
commands are not typically used in Dockerized CI pipelines.