diff/push: restore standard DELIMITER after each schema
Skeema v1.10.0 improved the output handling of `skeema diff` and `skeema
push` when manipulating multiple routines in a row, to avoid unnecessarily
switching between standard and nonstandard DELIMITERs (restoring the standard
";" delimiter and then immediately going back to "//" ahead of the next
routine) in commit 93ad4c4.

However, the new logic did not ensure that the standard delimiter was back in
use by the end of the command's output, which could be problematic for users
who pipe `skeema diff`'s STDOUT to another program and then add additional
output afterwards.

This commit improves the output by ensuring the standard semicolon delimiter
is back in use by the end of the output for each schema, which also inherently
ensures it is back in use by the end of the command's overall output.