diff/push: improve logging for unsafe statements
This commit makes several improvements to the log output of `skeema diff` and
`skeema push` when unsafe statements are present and have not been allowed
by --allow-unsafe. Behavior changes in this commit include:

* If there are multiple unsafe statements generated in the diff, all of them
  are now logged. Previously, only the first such statement was logged.

* Each unsafe statement now has a corresponding reason logged, providing more
  information about why the statement is considered unsafe.

* Linting now occurs even if unsafe statements are present. Previously, the
  linting step was skipped in this situation.

* The log message containing the unsafe SQL statement was previously glitchy
  if the SQL contained a % symbol, for example in a LIKE expression.

* The final summary log message regarding skipped operations now properly only
  counts statements that would have been emitted or executed. Previously, the
  count included differences that were suppressed, such as auto-increment
  discrepancies.

The overall goal of these changes is to make unsafe statement detection more
similar to linter error detection. Both steps are in the general category of
safety guardrails, and it is beneficial to present all such errors to the user
at one time, so that they can revise their schema repo change accordingly in
one step.