internal applier refactor: add Plan struct, improve tests
This commit refactors a chunk of the applier subpackage, with an overall goal
of starting to improve testability of this logic. Previously, due to tangled
logical dependencies, the applier subpackage was almost entirely relying on
coverage from integration tests in the main package for the Skeema command.
That's still the case here, but this commit starts to chip away at the
situation and improves coverage from applier's own tests.

The most important part of this refactor is the introduction of an
applier.Plan struct, and the ability to create a Plan being separate from
applying it.

No changes in external functionality are present in this commit.